requirements.txt
panda
pyarrow
tqdm
pytorch-lightning
optuna
transformers
line_profiler
matplotlib
xgboost
tsfresh
plotly
librosa
PyWavelets
timm